requirements.txt
# requirements.txt
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
google-cloud-bigquery
